UPDATE: create DateFormatterExtension + DoubleExtension and configure DataExtension
